pyproject.toml
[project]
name = "etl-data-fundation"
version = "0.1.0"
description = "Publicação dos dados silver para gold"
authors = [
    {name = "Duarte, Wilson Beirigo",email = "wilson.duarte@aecom.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "geopandas (>=1.0.1,<2.0.0)",
    "openpyxl (>=3.1.5,<4.0.0)",
    "geoserver-rest (>=2.9.1,<3.0.0)",
    "jinja2 (>=3.1.6,<4.0.0)",
    "pdfkit (>=1.0.0,<2.0.0)",
    "wkhtmltopdf (>=0.2,<0.3)",
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.3,<0.35.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "dotenv (>=0.9.9,<0.10.0)",
    "gdal (==3.8.4)",
    "certifi (>=2025.7.9,<2026.0.0)",
    "owslib (>=0.34.1,<0.35.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
